BE/pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pandas>=3.0.0",
    "pydantic[email]>=2.12.5",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.2.1",
]
